refactor(web): restructure src directory layout
- move CodemirrorEditor.vue from views/ to components/editor/
- move 10 dialog components into components/editor/dialogs/
- rename FolderSourcePanel/ to folder-source-panel/ (kebab-case)
- remove empty styles/ directory
- fix Pinia store IDs to camelCase (AIConfig -> aiConfig, AIImageConfig -> aiImageConfig)
- rename useQuickCommands to useQuickCommandsStore for consistency
