+++ apps/web/src/components/editor/dialogs/FormulaEditorDialog.vue
