+++ apps/web/src/components/editor/dialogs/LocalImageUploadDialog.vue
