+++ apps/web/src/components/editor/dialogs/MpAccountConfigDialog.vue
