+++ apps/web/src/components/editor/dialogs/UploadImgDialog.vue
