+++ apps/web/src/components/editor/folder-source-panel/index.vue
